+++ src/Investing/Investing.Application/Plans/Commands/DeletePlanCommand.cs
@@ -1,11 +1,15 @@
 ﻿using System;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using FluentValidation;
+using Hive.Common.Core;
 using Hive.Common.Core.Exceptions;
+using Hive.Common.Core.Interfaces;
 using Hive.Investing.Application.Interfaces;
 using Hive.Investing.Domain.Entities;
 using MediatR;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 
@@ -24,12 +28,14 @@ namespace Hive.Investing.Application.Plans.Commands
         }
     }
 
-    public class DeletePlanCommandHandler : IRequestHandler<DeletePlanCommand>
+    public class DeletePlanCommandHandler : AuthorizationRequestHandler<Plan>, IRequestHandler<DeletePlanCommand>
     {
         private readonly IInvestingDbContext _context;
         private readonly ILogger<DeletePlanCommandHandler> _logger;
 
-        public DeletePlanCommandHandler(IInvestingDbContext context, ILogger<DeletePlanCommandHandler> logger)
+        public DeletePlanCommandHandler(IInvestingDbContext context, ILogger<DeletePlanCommandHandler> logger,
+            ICurrentUserService currentUserService, IAuthorizationService authorizationService)
+            : base(currentUserService, authorizationService)
         {
             _context = context ?? throw new ArgumentNullException(nameof(context));
             _logger = logger ?? throw new ArgumentNullException(nameof(logger));
@@ -43,6 +49,13 @@ namespace Hive.Investing.Application.Plans.Commands
             {
                 _logger.LogWarning("Plan with id {Id} was not found", request.Id);
                 throw new NotFoundException(nameof(Plan), request.Id);
+            }
+            
+            var result = await base.AuthorizeAsync(plan,  new [] {"OnlyOwnerPolicy"});
+            
+            if (!result.All(s => s.Succeeded))
+            {
+                throw new ForbiddenAccessException();
             }
 
             _context.Plans.Remove(plan);
